Refine MCP resource lookup and split tests
